add tests for level hints
