+++ tracing-subscriber/tests/subscriber_filters/vec.rs
@@ -1,0 +1,117 @@
+use super::*;
+use tracing::Collect;
+
+#[test]
+fn with_filters_unboxed() {
+    let (trace_subscriber, trace_handle) = subscriber::named("trace")
+        .event(event::mock().at_level(Level::TRACE))
+        .event(event::mock().at_level(Level::DEBUG))
+        .event(event::mock().at_level(Level::INFO))
+        .done()
+        .run_with_handle();
+    let trace_subscriber = trace_subscriber.with_filter(LevelFilter::TRACE);
+
+    let (debug_subscriber, debug_handle) = subscriber::named("debug")
+        .event(event::mock().at_level(Level::DEBUG))
+        .event(event::mock().at_level(Level::INFO))
+        .done()
+        .run_with_handle();
+    let debug_subscriber = debug_subscriber.with_filter(LevelFilter::DEBUG);
+
+    let (info_subscriber, info_handle) = subscriber::named("info")
+        .event(event::mock().at_level(Level::INFO))
+        .done()
+        .run_with_handle();
+    let info_subscriber = info_subscriber.with_filter(LevelFilter::INFO);
+
+    let _collector = tracing_subscriber::registry()
+        .with(vec![trace_subscriber, debug_subscriber, info_subscriber])
+        .set_default();
+
+    tracing::trace!("hello trace");
+    tracing::debug!("hello debug");
+    tracing::info!("hello info");
+
+    trace_handle.assert_finished();
+    debug_handle.assert_finished();
+    info_handle.assert_finished();
+}
+
+#[test]
+fn with_filters_boxed() {
+    let (unfiltered_subscriber, unfiltered_handle) = subscriber::named("unfiltered")
+        .event(event::mock().at_level(Level::TRACE))
+        .event(event::mock().at_level(Level::DEBUG))
+        .event(event::mock().at_level(Level::INFO))
+        .done()
+        .run_with_handle();
+    let unfiltered_subscriber = unfiltered_subscriber.boxed();
+
+    let (debug_subscriber, debug_handle) = subscriber::named("debug")
+        .event(event::mock().at_level(Level::DEBUG))
+        .event(event::mock().at_level(Level::INFO))
+        .done()
+        .run_with_handle();
+    let debug_subscriber = debug_subscriber.with_filter(LevelFilter::DEBUG).boxed();
+
+    let (target_subscriber, target_handle) = subscriber::named("target")
+        .event(event::mock().at_level(Level::INFO))
+        .done()
+        .run_with_handle();
+    let target_subscriber = target_subscriber
+        .with_filter(filter::filter_fn(|meta| meta.target() == "my_target"))
+        .boxed();
+
+    let _collector = tracing_subscriber::registry()
+        .with(vec![
+            unfiltered_subscriber,
+            debug_subscriber,
+            target_subscriber,
+        ])
+        .set_default();
+
+    tracing::trace!("hello trace");
+    tracing::debug!("hello debug");
+    tracing::info!(target: "my_target", "hello my target");
+
+    unfiltered_handle.assert_finished();
+    debug_handle.assert_finished();
+    target_handle.assert_finished();
+}
+
+#[test]
+fn mixed_max_level_hint() {
+    let unfiltered = subscriber::named("unfiltered").run().boxed();
+    let info = subscriber::named("info")
+        .run()
+        .with_filter(LevelFilter::INFO)
+        .boxed();
+    let debug = subscriber::named("debug")
+        .run()
+        .with_filter(LevelFilter::DEBUG)
+        .boxed();
+
+    let collector = tracing_subscriber::registry().with(vec![unfiltered, info, debug]);
+
+    assert_eq!(collector.max_level_hint(), None);
+}
+
+#[test]
+fn all_filtered_max_level_hint() {
+    let warn = subscriber::named("warn")
+        .run()
+        .with_filter(LevelFilter::WARN)
+        .boxed();
+    let info = subscriber::named("info")
+        .run()
+        .with_filter(LevelFilter::INFO)
+        .boxed();
+    let debug = subscriber::named("debug")
+        .run()
+        .with_filter(LevelFilter::DEBUG)
+        .boxed();
+
+    let collector = tracing_subscriber::registry().with(vec![warn, info, debug]);
+
+    assert_eq!(collector.max_level_hint(), Some(LevelFilter::DEBUG));
+}
